[RF2] Move logging and specification into RF2Context.java abstract class
