+++ src/main/java/com/b2international/rf2/RF2Context.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2019 B2i Healthcare Pte Ltd, http://b2i.sg
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.b2international.rf2;
+
+import com.b2international.rf2.spec.RF2Specification;
+
+/**
+ * @since 0.3
+ */
+public abstract class RF2Context {
+
+    protected final RF2Specification specification;
+    protected final Console console;
+
+    public RF2Context(RF2Specification specification, Console console) {
+        this.specification = specification;
+        this.console = console;
+    }
+
+    public void log(String message, Object... args) {
+        console.log(message, args);
+    }
+
+    public void warn(String message, Object... args) {
+        console.warn(message, args);
+    }
+
+    public void error(String message, Object... args) {
+        console.error(message, args);
+    }
+
+    public RF2Specification getSpecification() {
+        return specification;
+    }
+
+}
