+++ src/main/java/com/b2international/rf2/RF2CreateContext.java
@@ -29,38 +29,19 @@ import com.b2international.rf2.spec.RF2Specification;
 /**
  * @since 0.1
  */
-public final class RF2CreateContext {
+public final class RF2CreateContext extends RF2Context{
 
-	private final RF2Specification specification;
 	private final List<RF2File> sources;
-	private final Console log;
-	
+
 	public RF2CreateContext(RF2Specification specification, List<RF2File> sources, Console log) {
-		this.specification = specification;
-		this.log = log;
+		super(specification, log);
 		this.sources = sources == null ? Collections.emptyList() : sources;
-	}
-	
-	public RF2Specification getSpecification() {
-		return specification;
 	}
 	
 	public List<RF2File> getSources() {
 		return sources;
 	}
-	
-	public void log(String message, Object... args) {
-		log.log(message, args);
-	}
 
-	public void warn(String message, Object... args) {
-		log.warn(message, args);
-	}
-
-	public void error(String message, Object... args) {
-		log.error(message, args);
-	}
-	
 	public void visitSourceRows(Predicate<RF2ContentFile> fileFilter, Predicate<String[]> lineFilter, boolean parallel, Consumer<String[]> visitor) throws IOException {
 		for (RF2File source : sources) {
 			source.visit(file -> {
